Modules/Messenger/bin/verifyAppSize/requirements.txt
PyYAML==5.3.0
requests==2.24.0